requirements.txt
slack-bolt>=1.18.0
google-generativeai>=0.3.0
pandas>=2.0.0
sqlalchemy>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
rapidfuzz>=3.0.0
Faker>=20.0.0
